+++ CHANGELOG.md
@@ -10,7 +10,21 @@
 - No longer initializes color mode based on `prefers-color-scheme` media query - TODO: provide alternative API
 - Removes layout components (`Layout`, `Header`, `Main`, `Footer`)
 
+- `@theme-ui/match-media`: add option for default index in hook #460
+
+## v0.2.51 2019-12-03
+
+- `@theme-ui/editor`: fix color picker #498
+
+## v0.2.50 2019-12-02
+
+- `@theme-ui/components`: fix NavLink base styles #497
+
+## v0.2.49 2019-11-15
+
+- `@theme-ui/components`: add more components #458
 - `@theme-ui/color`: add `alpha` utility #441
+- `@theme-ui/match-media`: Add default breakpoint index argument for SSR
 
 ## v0.2.48 2019-11-07